server/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.41
alembic==1.15.2
pydantic==2.11.4
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.1.0
jinja2==3.1.6
